#MNDS-25

+++ nds-client/src/main/java/br/com/abril/nds/controllers/cadastro/ProdutoEdicaoController.java
@@ -208,12 +208,12 @@ public class ProdutoEdicaoController extends BaseController {
 	@Post
 	@Path("/pesquisarEdicoes.json")
 	public void pesquisarEdicoes(FiltroProdutoDTO filtro,
-			Date dataLancamentoDe, Date dataLancamentoAte, BigDecimal precoDe,
-			BigDecimal precoAte , StatusLancamento situacaoLancamento,
+			Date dataLancamentoDe, Date dataLancamentoAte, Double precoDe,
+			Double precoAte , StatusLancamento situacaoLancamento,
 			String codigoDeBarras, boolean brinde, ModoTela modoTela,
             String sortorder, String sortname, int page, int rp) {
 		
-		Intervalo<BigDecimal> intervaloPreco = null;
+		Intervalo<Double> intervaloPreco = null;
 		Intervalo<Date> intervaloLancamento = null;
 		
 		// Validar:
@@ -233,7 +233,7 @@ public class ProdutoEdicaoController extends BaseController {
 			if(precoDe.compareTo(precoAte) > 0){
 				throw new ValidacaoException(TipoMensagem.WARNING, "Por favor, preencha o intervalo válido de 'Preço'!");
 			}
-			intervaloPreco = new Intervalo<BigDecimal>(precoDe, precoAte);
+			intervaloPreco = new Intervalo<Double>(precoDe, precoAte);
 		}	
 	
 		// Pesquisar:
